refactor(config): localize runtime policy validation helpers by domain
