+++ crates/config/src/runtime/policies/admission.rs
@@ -1,9 +1,7 @@
 use std::{collections::HashMap, time::Duration};
 
 use super::{
-    config_invalid, is_valid_connect_authority, is_valid_http_token, is_valid_request_key_spec,
-    normalize_nonempty_string_vec, normalize_optional_string, normalize_string_vec,
-    require_nonzero_usize,
+    config_invalid, normalize_optional_string,
 };
 use crate::{
     config::Resilience,
@@ -17,6 +15,82 @@ use super::{
     },
     watchdog::normalize_watchdog_policy,
 };
+
+fn require_nonzero_usize(name: &str, value: usize) -> Result<(), RuntimeConfigError> {
+    if value == 0 {
+        return Err(config_invalid(format!("{name} must be greater than 0")));
+    }
+    Ok(())
+}
+
+fn normalize_string_vec(values: &[String]) -> Vec<String> {
+    values
+        .iter()
+        .map(|value| value.trim())
+        .filter(|value| !value.is_empty())
+        .map(ToOwned::to_owned)
+        .collect()
+}
+
+fn normalize_nonempty_string_vec(
+    field_name: &str,
+    values: &[String],
+) -> Result<Vec<String>, RuntimeConfigError> {
+    let normalized = normalize_string_vec(values);
+    if normalized.len() != values.len() {
+        return Err(config_invalid(format!(
+            "{field_name} must not contain empty values"
+        )));
+    }
+    Ok(normalized)
+}
+
+fn is_valid_http_token(value: &str) -> bool {
+    !value.is_empty()
+        && value.bytes().all(|byte| {
+            matches!(
+                byte,
+                b'!'
+                    | b'#'..=b'\''
+                    | b'*'
+                    | b'+'
+                    | b'-'
+                    | b'.'
+                    | b'0'..=b'9'
+                    | b'A'..=b'Z'
+                    | b'^'
+                    | b'_'
+                    | b'`'
+                    | b'a'..=b'z'
+                    | b'|'
+                    | b'~'
+            )
+        })
+}
+
+fn is_valid_connect_authority(authority: &str) -> bool {
+    let Some((host, port)) = authority.rsplit_once(':') else {
+        return false;
+    };
+    !host.trim().is_empty() && port.parse::<u16>().ok().is_some_and(|parsed| parsed > 0)
+}
+
+fn is_valid_request_key_spec(key_spec: &str) -> bool {
+    let key_spec = key_spec.trim().to_ascii_lowercase();
+    matches!(
+        key_spec.as_str(),
+        "path"
+            | "authority"
+            | "method"
+            | "cid"
+            | "sticky-cid"
+            | "peer_ip"
+            | "client_ip"
+            | "bearer_token"
+    ) || key_spec.split_once(':').is_some_and(|(source, key_name)| {
+        !key_name.trim().is_empty() && matches!(source.trim(), "header" | "cookie" | "query")
+    })
+}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RuntimeScopedRateLimitPolicy {
