+++ crates/config/src/runtime/policies/timeouts.rs
@@ -1,10 +1,17 @@
 use std::time::Duration;
 
-use super::{config_invalid, require_nonzero_u64};
+use super::config_invalid;
 use crate::{
     config::Performance,
     runtime::RuntimeConfigError,
 };
+
+fn require_nonzero_u64(name: &str, value: u64) -> Result<(), RuntimeConfigError> {
+    if value == 0 {
+        return Err(config_invalid(format!("{name} must be greater than 0")));
+    }
+    Ok(())
+}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct RuntimeTimeoutPolicy {
